# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "CSM Deployment API",
        "title": "CSM Deployment API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/configuration-files": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "List all configuration files",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "configuration-file"
                ],
                "summary": "List all configuration files",
                "operationId": "list-config-file",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Name of the configuration file",
                        "name": "config_name",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/ConfigFileResponse"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Create a new configuration file",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "configuration-file"
                ],
                "summary": "Create a new configuration file",
                "operationId": "create-config-file",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Name of the configuration file",
                        "name": "name",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "Configuration file",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/ConfigFileResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    }
                }
            }
        },
        "/configuration-files/{id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Get a configuration file",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "configuration-file"
                ],
                "summary": "Get a configuration file",
                "operationId": "get-config-file",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Configuration file ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ConfigFileResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Delete a configuration file",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "configuration-file"
                ],
                "summary": "Delete a configuration file",
                "operationId": "delete-config-file",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Configuration file ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": ""
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Update a configuration file",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "configuration-file"
                ],
                "summary": "Update a configuration file",
                "operationId": "update-config-file",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Configuration file ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Name of the configuration file",
                        "name": "name",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "Configuration file",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "ConfigFileResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "ErrorMessage": {
            "type": "object",
            "properties": {
                "arguments": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "code": {
                    "description": "HTTPStatusEnum  Possible HTTP status values of completed or failed jobs",
                    "type": "integer",
                    "enum": [
                        200,
                        201,
                        202,
                        204,
                        400,
                        401,
                        403,
                        404,
                        422,
                        429,
                        500,
                        503
                    ]
                },
                "message": {
                    "description": "Message string.",
                    "type": "string"
                },
                "message_l10n": {
                    "description": "Localized message",
                    "type": "object"
                },
                "severity": {
                    "description": "SeverityEnum - The severity of the condition\n* INFO - Information that may be of use in understanding the failure. It is not a problem to fix.\n* WARNING - A condition that isn't a failure, but may be unexpected or a contributing factor. It may be necessary to fix the condition to successfully retry the request.\n* ERROR - An actual failure condition through which the request could not continue.\n* CRITICAL - A failure with significant impact to the system. Normally failed commands roll back and are just ERROR, but this is possible",
                    "type": "string",
                    "enum": [
                        "INFO",
                        "WARNING",
                        "ERROR",
                        "CRITICAL"
                    ]
                }
            }
        },
        "ErrorResponse": {
            "type": "object",
            "properties": {
                "http_status_code": {
                    "description": "HTTPStatusEnum  Possible HTTP status values of completed or failed jobs",
                    "type": "integer",
                    "enum": [
                        200,
                        201,
                        202,
                        204,
                        400,
                        401,
                        403,
                        404,
                        422,
                        429,
                        500,
                        503
                    ]
                },
                "messages": {
                    "description": "A list of messages describing the failure encountered by this request. At least one will\nbe of Error severity because Info and Warning conditions do not cause the request to fail",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ErrorMessage"
                    }
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "BasicAuth": {
            "type": "basic"
        }
    }
}